requirements.txt
slacker==0.9.60
websocket-client==0.44.0

